Start sandbox in-proc
